requirements.txt
streamlit
opencv-python
mediapipe
openai
python-dotenv
pyyaml
numpy